predictor/requirements.txt
xgboost
pandas
numpy
huggingface_hub
psycopg2-binary
sentry-sdk
